Update NUMBER.MaxRateValue.md
The decimal separator used is actually a comma not a point. The code handles an input using a point as the separator and changes the separator to a comma: e.g. 0.05 input to 0,05.

I also updated the description of the influence of MaxRateType and included a use case example in the last paragraph.